Add test cases for every kind of namespace support
