+++ test/namespace-none/generate.go
@@ -1,0 +1,3 @@
+package avro
+
+//go:generate $GOPATH/bin/gogen-avro --namespaced-names none --containers . namespace.avsc
