+++ test/namespace-short/generate.go
@@ -1,0 +1,3 @@
+package avro
+
+//go:generate $GOPATH/bin/gogen-avro --namespaced-names short --containers . namespace.avsc
