--- test/namespace/generate.go
@@ -1,3 +1,0 @@
-package avro
-
-//go:generate $GOPATH/bin/gogen-avro --namespaced-names --containers . namespace.avsc
